Move node model helpers into inventory module
