+++ custom_components/termoweb/nodes.py
@@ -1,4 +1,4 @@
-"""Node model abstractions and helpers for TermoWeb devices."""
+"""Node helpers for TermoWeb devices."""
 
 from __future__ import annotations
 
@@ -7,76 +7,18 @@ import logging
 from typing import Any, cast
 
 from .const import DOMAIN
+from .inventory import (
+    AccumulatorNode,
+    HeaterNode,
+    NODE_CLASS_BY_TYPE,
+    Node,
+    PowerMonitorNode,
+    ThermostatNode,
+    normalize_node_addr,
+    normalize_node_type,
+)
 
 _LOGGER = logging.getLogger(__name__)
-
-
-def _normalize_node_identifier(
-    value: Any,
-    *,
-    default: str = "",
-    use_default_when_falsey: bool = False,
-    lowercase: bool,
-) -> str:
-    """Return ``value`` as a normalised node identifier string."""
-
-    raw = value
-    if use_default_when_falsey and not raw:
-        raw = default
-
-    try:
-        normalized = str(raw).strip()
-    except Exception:  # pragma: no cover - defensive  # noqa: BLE001
-        normalized = ""
-    else:
-        if lowercase:
-            normalized = normalized.lower()
-
-    if normalized:
-        return normalized
-
-    if default and not use_default_when_falsey:
-        try:
-            normalized_default = str(default).strip()
-        except Exception:  # pragma: no cover - defensive  # noqa: BLE001
-            return ""
-        if lowercase:
-            normalized_default = normalized_default.lower()
-        return normalized_default
-
-    return ""
-
-
-def normalize_node_type(
-    value: Any,
-    *,
-    default: str = "",
-    use_default_when_falsey: bool = False,
-) -> str:
-    """Return ``value`` as a normalised node type string."""
-
-    return _normalize_node_identifier(
-        value,
-        default=default,
-        use_default_when_falsey=use_default_when_falsey,
-        lowercase=True,
-    )
-
-
-def normalize_node_addr(
-    value: Any,
-    *,
-    default: str = "",
-    use_default_when_falsey: bool = False,
-) -> str:
-    """Return ``value`` as a normalised node address string."""
-
-    return _normalize_node_identifier(
-        value,
-        default=default,
-        use_default_when_falsey=use_default_when_falsey,
-        lowercase=False,
-    )
 
 
 HEATER_NODE_TYPES: frozenset[str] = frozenset({"htr", "acm"})
@@ -215,139 +157,6 @@ def _extract_snapshot_name(payloads: Iterable[Mapping[str, Any]]) -> str:
     return ""
 
 
-class Node:
-    """Base representation of a TermoWeb node."""
-
-    __slots__ = ("_node_name", "addr", "type")
-    NODE_TYPE = ""
-
-    def __init__(
-        self,
-        *,
-        name: str | None,
-        addr: str | int,
-        node_type: str | None = None,
-    ) -> None:
-        """Initialise a node with normalised metadata."""
-
-        resolved_type = normalize_node_type(
-            node_type,
-            default=self.NODE_TYPE,
-            use_default_when_falsey=True,
-        )
-        if not resolved_type:
-            msg = "node_type must be provided"
-            raise ValueError(msg)
-
-        addr_str = normalize_node_addr(addr)
-        if not addr_str:
-            msg = "addr must be provided"
-            raise ValueError(msg)
-
-        self.addr = addr_str
-        self.type = resolved_type
-        self._node_name = ""
-        self.name = name if name is not None else ""
-
-    @property
-    def name(self) -> str:
-        """Return the friendly name for the node."""
-
-        attr_name = getattr(self, "_attr_name", None)
-        if isinstance(attr_name, str) and attr_name.strip():
-            return attr_name
-        return self._node_name
-
-    @name.setter
-    def name(self, value: str | None) -> None:
-        """Update the stored friendly name for the node."""
-        cleaned = str(value or "").strip()
-        self._node_name = cleaned
-        if hasattr(self, "_attr_name"):
-            self._attr_name = cleaned
-
-    def as_dict(self) -> dict[str, Any]:
-        """Return a serialisable snapshot of core node metadata."""
-
-        return {
-            "name": self.name,
-            "addr": self.addr,
-            "type": self.type,
-        }
-
-
-NodeDescriptor = Node | tuple[str, str | int]
-
-
-class HeaterNode(Node):
-    """Heater node (type ``htr``)."""
-
-    __slots__ = ()
-    NODE_TYPE = "htr"
-
-    def __init__(self, *, name: str | None, addr: str | int) -> None:
-        """Initialise a heater node."""
-
-        super().__init__(name=name, addr=addr)
-
-    def supports_boost(self) -> bool:
-        """Return whether the node natively exposes boost/runback control."""
-
-        return False
-
-
-class AccumulatorNode(HeaterNode):
-    """Storage heater / accumulator node (type ``acm``)."""
-
-    __slots__ = ()
-    NODE_TYPE = "acm"
-
-    def supports_boost(self) -> bool:
-        """Return whether the accumulator exposes boost/runback."""
-
-        return True
-
-
-class PowerMonitorNode(Node):
-    """Power monitor node (type ``pmo``)."""
-
-    __slots__ = ()
-    NODE_TYPE = "pmo"
-
-    def __init__(self, *, name: str | None, addr: str | int) -> None:
-        """Initialise a power monitor node."""
-
-        super().__init__(name=name, addr=addr)
-
-    def power_level(self) -> float:
-        """Return the reported power level (stub)."""
-
-        raise NotImplementedError
-
-
-class ThermostatNode(Node):
-    """Thermostat node (type ``thm``)."""
-
-    __slots__ = ()
-    NODE_TYPE = "thm"
-
-    def __init__(self, *, name: str | None, addr: str | int) -> None:
-        """Initialise a thermostat node."""
-
-        super().__init__(name=name, addr=addr)
-
-    def capabilities(self) -> dict[str, Any]:
-        """Return thermostat capabilities (stub)."""
-
-        raise NotImplementedError
-
-
-NODE_CLASS_BY_TYPE: dict[str, type[Node]] = {
-    HeaterNode.NODE_TYPE: HeaterNode,
-    AccumulatorNode.NODE_TYPE: AccumulatorNode,
-    PowerMonitorNode.NODE_TYPE: PowerMonitorNode,
-    ThermostatNode.NODE_TYPE: ThermostatNode,
-}
 
 
 def _iter_node_payload(raw_nodes: Any) -> Iterable[dict[str, Any]]:
